+++ yumyum_front/src/views/Authentication/Login/Login.tsx
@@ -9,13 +9,12 @@ import * as css from "./Style";
 import axios from "axios";
 import {
   AUTH_PATH_SIGN_UP,
-  AUTH_PATH_LOGIN,
   MAIN_PATH,
 } from "../../../constants";
 import { useCookies } from "react-cookie";
 import useAuthStore from "../../../Stroes/auth.store";
 
-export default function AuthUser() {
+export default function Login() {
   const navigate = useNavigate();
 
   const [userLogInInfo, setUserLogInInfo] = useState<UserLogInInfo>({
@@ -25,11 +24,7 @@ export default function AuthUser() {
 
   const [error, setError] = useState<string>("");
   const [, setCookies] = useCookies(["token"]);
-  const { login, logout, user } = useAuthStore();
-
-  useEffect(() => {
-    console.log("로그인 상태", user)
-  }, [user])
+  const { login } = useAuthStore();
 
   const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     e.preventDefault();
@@ -42,34 +37,26 @@ export default function AuthUser() {
 
   const handleSubmit = async (e: React.MouseEvent<HTMLButtonElement>) => {
     e.preventDefault();
-    if (!user) {
-      try {
-        const response = await axios.post(
-          `http://localhost:4041/api/v1/auth/login`,
-          userLogInInfo
-        );
+    try {
+      const response = await axios.post(
+        `http://localhost:4041/api/v1/auth/login`,
+        userLogInInfo
+      );
+      console.log("로그인 응답:", response.data.data);
 
-        if (response.data) {
-          signInSuccessResponse(response.data.data);
-          console.log(response.data.data);
-        }
-      } catch (e) {
-        setError("로그인 중 문제가 발생했습니다.");
-      }
-    } else {
-        // console.log("first")
-        setCookies("token", "");
-        logout();
-        navigate(AUTH_PATH_LOGIN);
+      if (response.data && response.data.data) {
+        signInSuccessResponse(response.data.data);
+        console.log(response.data.data);
+      } 
+    } catch (e) {
+      setError("로그인 중 문제가 발생했습니다.");
     }
   };
 
   const setToken = (token: string, exprTime: number) => {
     const expires = new Date(Date.now() + exprTime);
-    setCookies("token", token, {
-      path: "/",
-      expires,
-    });
+    console.log("토큰", token, "만료시간", expires)
+    setCookies("token", token, { path: "/", expires });
   };
 
   const signInSuccessResponse = (data: SignInResponseDto) => {
@@ -89,11 +76,7 @@ export default function AuthUser() {
     }
   };
 
-  const handleLogout = () => {
-    setCookies("token", "", { path: "/", expires: new Date(0) }); 
-    logout();
-    console.log("로그아웃 성공");
-  }
+
   return (
     <>
       <h2 css={css.logInTitle}>로그인</h2>
@@ -138,11 +121,10 @@ export default function AuthUser() {
           <Button
             css={css.submitButton}
             type="submit"
-            onClick={user ? handleLogout : handleSubmit}
+            onClick={handleSubmit}
             variant="contained"
             color="primary"
-          >
-            {user ? "로그아웃" : "로그인"}
+          >로그인
           </Button>
         </Box>
         <Box css={css.link}>
